requirements.txt
discord.py==2.4.0
pyyaml==6.0.2
